[Rename] 파일 구성 수정
