+++ src/components/root/login/Login.js
@@ -1,12 +1,12 @@
 import logoIcon from '@assets/logo-icon.png';
-import '@styles/login/Login.scss';
+import '@styles/root/login/Login.scss';
 
-import LoginForm from './login/LoginForm';
-import LoginNavigation from './login/LoginNavigation';
+import LoginForm from './LoginForm';
+import LoginNavigation from './LoginNavigation';
 
 const Login = ({ error, isSubmitting }) => {
   return (
-    <section>
+    <section className='login-section'>
       <img className='logo' src={logoIcon} alt='logo icon' />
       <h3 className='description'>여행의 재미를 더하는 SNS</h3>
