+++ src/pages/logout/Logout.js
