+++ src/pages/social/SocialLogin.js
