+++ src/styles/root/login/Login.scss
@@ -1,4 +1,4 @@
-section {
+.login-section {
   display: flex;
   flex-direction: column;
   align-items: center;
